changes to module naming - Issue #23
