--- src/js/jqueryModule.js
@@ -1,10 +1,0 @@
-const jqueryModule = (function() {
-  const message = "Loaded jQuery module";
-  console.log(message);
-  
-  return {
-    message: message
-  };
-})();
-
-module.exports = jqueryModule;
