+++ src/scss/_variables.scss
@@ -2,6 +2,6 @@
 $font-stack: 'Quicksand', sans-serif;
 
 // COLORS
-$background-start: #cfd9df;
-$background-end: #e2ebf0;
+$background-start: #ff9a9e;
+$background-end: #fad0c4;
 $gray: #c1c1c1;
